requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.17.1
